requirements.txt
altair==4.2.0
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.6.15.2
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
fonttools==4.37.1
gitdb==4.0.9
GitPython==3.1.27
idna==3.4
importlib-metadata==4.12.0
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.16.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
nltk==3.7
numpy==1.23.3
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
plotly==5.10.0
protobuf==3.20.2
pyarrow==9.0.0
pydeck==0.8.0b3
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.9.13
requests==2.28.1
rich==12.5.1
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
semver==2.13.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
streamlit==1.12.0
tenacity==8.0.1
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wordcloud==1.8.2.2
zipp==3.8.1
